Share homepage scroll progress
